% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eou_MakeADFun.R
\name{eou_MakeADFun}
\alias{eou_MakeADFun}
\title{Construct a \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} object for the exponential Ornstein-Uhlenbeck stochastic volatility model.}
\usage{
eou_MakeADFun(
  Xt,
  dt,
  log_Vt,
  alpha,
  log_gamma,
  mu,
  log_sigma,
  logit_rho,
  par_list
)
}
\arguments{
\item{Xt}{Vector of \code{nobs} asset log prices.}

\item{dt}{Interobservation time.}

\item{log_Vt}{Optional vector of \code{nobs} volatilities on the log standard deviation scale.  See 'Details'.}

\item{alpha}{Optional asset growth rate parameter.  See 'Details'.}

\item{log_gamma}{Optional log-volatility mean reversion parameter on the log scale.  See 'Details'.}

\item{mu}{Optional log-volatility mean parameter.  See 'Details'.}

\item{log_sigma}{Optional log-volatility diffusion parameter on the log scale.  See 'Details'.}

\item{logit_rho}{Optional correlation parameter between asset and volatility innovations, on the logit scale.  See 'Details'.}

\item{par_list}{Optional list with named elements consisting of a subset of \code{log_Vt}, \code{alpha}, \code{log_gamma}, \code{mu}, \code{log_sigma}, and \code{logit_rho}.  Values in \code{par_list} will supercede those of the corresponding individual argument if both are provided.}
}
\value{
The result of a call to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
\description{
Construct a \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} object for the exponential Ornstein-Uhlenbeck stochastic volatility model.
}
\details{
The exponential Ornstein-Uhlenbeck (eOU) stochastic volatility model for a single asset is given by the stochastic differential equation (SDE) ...

\code{eou_MakeADFun} implements the Euler approximation to this SDE...

The optional inputs \code{log_Vt}, \code{alpha}, ..., \code{logit_rho} can be set to initialize optimization routines.  The default values are \code{alpha = 0}, ..., \code{logit_rho = 0}, and \code{log_Vt} as the log of windowed standard deviation estimates returned by \code{\link[=sv_init]{sv_init()}}.

\code{eou_MakeADFun} is a wrapper to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.  This function may be called on the underlying C++ template provided by \pkg{svcommon} via\preformatted{TMB::MakeADFun(data = list(model = "sv_eou", ...),
               parameters = list(...),
               DLL = "svcommon_TMBExports",
               ...)
}
}
