% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eou_sim.R
\name{eou_sim}
\alias{eou_sim}
\title{Simulate time series from the exponential Ornstein-Uhlenbeck stochastic volatility model.}
\usage{
eou_sim(nobs, dt, X0, log_V0, alpha, log_gamma, mu, log_sigma, logit_rho, dB)
}
\arguments{
\item{nobs}{Length of time series.}

\item{dt}{Interobservation time.}

\item{X0}{Scalar or vector of \code{nseries} asset log prices at time \code{t = 0}.}

\item{log_V0}{Scalar or vector of \code{nseries} volatilities at time \code{t = 0}, on the log standard deviation scale.}

\item{alpha}{Scalar or vector of \code{nseries} growth rate parameters.}

\item{log_gamma}{Scalar or vector of \code{nseries} log-volatility mean reversion parameters on the log scale.}

\item{mu}{Scalar or vector of \code{nseries} log-volatility mean parameters.}

\item{log_sigma}{Scalar or vector of \code{nseries} log-volatility diffusion parameters on the log scale.}

\item{logit_rho}{Scalar or vector of \code{nseries} correlation parameters between asset and volatility innovations, on the logit scale.}

\item{dBt}{An optional list with elements \code{V} and \code{Z} corresponding to matrices of size \verb{nobs x nseries} of pre-specified Brownian innovations.  If missing these consist of iid draws from an \code{N(0, dt)} distribution.}
}
\value{
A list containing matrices \code{Xt} and \code{log_Vt} of \verb{nobs x nseries} of eOU observations, where each column corresponds to a process observed at times \verb{t = dt, 2dt, ..., nobs*dt}.
}
\description{
Simulate time series from the exponential Ornstein-Uhlenbeck stochastic volatility model.
}
