% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou_fit.R
\name{ou_fit}
\alias{ou_fit}
\title{Estimate parameters of an Ornstein-Uhlenbeck process.}
\usage{
ou_fit(Xt, dt)
}
\arguments{
\item{Xt}{Vector of log-prices.}

\item{dt}{Interobservation time.}
}
\value{
A named vector containing the maximum likelihood estimate of the OU parameters.
}
\description{
Calculate the MLE using least-squares on the one-step Euler approximation.
}
\details{
The Ornstein-Uhlenbeck process is described by the stochastic differential equation\preformatted{dXt = - gamma (Xt - mu) dt + sigma dBt.
}

Here, the MLE is computed analytically from the Euler approximation\preformatted{X_\{t+1\} | X_t ~ N(X_t - gamma (X_t - mu) dt, sigma^2 dt).
}
}
