% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp500.R
\docType{data}
\name{snp500}
\alias{snp500}
\title{Daily observations of S&P500 constituents.}
\format{
An object of class \code{data.frame} with 3408 rows and 48 columns.
}
\usage{
snp500
}
\description{
Daily observations of S&P500 constituents.
}
\keyword{datasets}
