% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_init.R
\name{sv_init}
\alias{sv_init}
\title{Initialize latent volatilities in a stochastic volatility model.}
\usage{
sv_init(Xt, dt, block_size)
}
\arguments{
\item{Xt}{Vector of log-prices.}

\item{dt}{Interobservation time.}

\item{block_size}{Block size over which to calculate variance.}
}
\value{
A vector of volatilities on the standard deviation scale, of the same length as \code{Xt}.
}
\description{
Initialize latent volatilities in a stochastic volatility model.
}
