% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svc_dB.R
\name{svc_dB}
\alias{svc_dB}
\title{Extract the Brownian increments from the SVC model.}
\usage{
svc_dB(Xt, log_VPt, log_Vt, dt, alpha, log_gamma, mu, log_sigma, logit_rho)
}
\arguments{
\item{Xt}{Matrix of \verb{nobs x (nasset + 1)} asset log prices, where the first column is that of the asset common-factor proxy.}

\item{log_VPt}{Vector of \code{nobs} volatility proxy values on the log standard deviation scale.  See 'Details'.}

\item{log_Vt}{Optional vector of \verb{nobs x (nasset + 1)} volatilities on the log standard deviation scale.  See 'Details'.}

\item{dt}{Interobservation time.}

\item{alpha}{Optional vector of \code{(nasset + 1)} asset growth rate parameters.  See 'Details'.}

\item{log_gamma}{Optional vector of \code{(nasset + 2)} log-volatility mean reversion parameters on the log scale.  The first two correspond to the volatility proxy and the common-factor asset's volatility, respectively.  See 'Details'.}

\item{mu}{Optional vector of \code{(nasset + 2)} log-volatility mean parameters.  See 'Details'.}

\item{log_sigma}{Optional vector of \code{(nasset + 2)} log-volatility diffusion parameters on the log scale.  See 'Details'.}

\item{logit_rho}{Optional vector of \code{(nasset + 1)} correlation parameters between asset and volatility innovations, on the logit scale.  The first one is that of the common-factor asset proxy.  See 'Details'.}
}
\value{
A list with elements:
\describe{
\item{\code{V}}{An \verb{nobs x (nasset+2)} matrix of the log-volatility innovations.}
\item{\code{X}}{An \verb{nobs x (nasset+1)} matrix of log-asset innovations.}
\item{\code{Z}}{An \verb{nobs x (nasset+1)} matrix of residual log-asset innovations, after account for the log-volatility innovations.  That is,
\verb{dB_Z = (dB_X - rho dB_V) / sqrt(1 - rho^2).}
}
}
}
\description{
Extract the Brownian increments from the SVC model.
}
