% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svc_sim.R
\name{svc_sim}
\alias{svc_sim}
\title{Simulate time series from the SVC stochastic volatility model.}
\usage{
svc_sim(
  nobs,
  dt,
  X0,
  log_VP0,
  log_V0,
  alpha,
  log_gamma,
  mu,
  log_sigma,
  logit_rho,
  logit_tau,
  logit_omega,
  dBt
)
}
\arguments{
\item{nobs}{Length of time series.}

\item{dt}{Interobservation time.}

\item{X0}{Vector of length \code{nasset + 1} or matrix of size \verb{(nasset + 1) x nseries} of asset log prices at time \code{t = 0}.}

\item{log_VP0}{Scalar or vector of length \code{nseries} of volatility proxy values at time \code{t = 0} on the log standard deviation scale.}

\item{log_V0}{Vector of length \code{nasset + 1} or matrix of size \verb{(nasset + 1) x nseries} of volatilities at time \code{t = 0} on the log standard deviation scale.}

\item{alpha}{Vector of length \code{(nasset + 1)} or matrix of size \verb{(nasset + 1) x nseries} of asset growth rate parameters.}

\item{log_gamma}{Vector of length \code{nasset + 2} or matrix of size \verb{(nasset + 2) x nseries} log-volatility mean reversion parameters on the log scale.  The first two correspond to the volatility proxy and the common-factor asset's volatility, respectively.}

\item{mu}{Vector of length \code{nasset + 2} or matrix of size \verb{(nasset + 2) x nseries} of log-volatility mean parameters.}

\item{log_sigma}{Vector of length \code{nasset + 2} or matrix of size \verb{(nasset + 2) x nseries} or log-volatility diffusion parameters on the log scale.}

\item{logit_rho}{Vector of length \code{nasset + 1} or matrix of size \verb{(nasset + 1) x nseries} correlation parameters between asset and volatility innovations, on the logit scale.  The first one is that of the common-factor asset proxy.  See 'Details'.}

\item{logit_tau}{Vector of length \code{nasset + 1} or matrix of size \verb{(nasset + 1) x nseries} correlation parameters between the latent volatilities and the volatility proxy.}

\item{logit_omega}{Vector of length \code{nasset} or matrix of size \verb{nasset x nseries} correlation parameters between the residual asset price of the common-factor proxy and the other residual asset prices.}

\item{dBt}{An optional list of Brownian innovations:
\describe{
\item{\code{VP}}{An \verb{nobs x nseries} matrix.}
\item{\code{V}}{An \verb{nobs x (nasset+1) x nseries} array.}
\item{\code{Z}}{An \verb{nobs x (nasset+1) x nseries} array.}
}
If missing these are all iid draws from an \code{N(0, dt)} distribution.}
}
\value{
A list containing arrays \code{Xt} and \code{log_Vt} of size \verb{nobs x (nasset + 1) x nseries}, and the matrix \code{log_VPt} of size \verb{nobs x nseries} containing the simulations of \code{nseries} SVC processes observed at times \verb{t = dt, 2dt, ..., nobs*dt}.
}
\description{
Simulate time series from the SVC stochastic volatility model.
}
