% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svc_update.R
\name{svc_update}
\alias{svc_update}
\title{Update the parameters and latent variables of an SVC model.}
\usage{
svc_update(fit, old_par, iasset)
}
\arguments{
\item{fit}{An object returned by \code{\link[=svc_MakeADFun]{svc_MakeADFun()}} or \code{\link[=eou_MakeADFun]{eou_MakeADFun()}}, typically after it has been used in a numerical optimization routine.}

\item{old_par}{A list of SVC parameters to be updated.  Must be in the same format as the \code{par_list} argument of \code{\link[=svc_MakeADFun]{svc_MakeADFun()}}.}

\item{iasset}{Which parameter(s) to update.  See corresponding argument of \code{\link[=svc_MakeADFun]{svc_MakeADFun()}}.}
}
\value{
A list in the same format as \code{old_par} but with the corresponding parameters and latent variables updated with those of \code{svc_fit}.
}
\description{
Update the parameters and latent variables of an SVC model.
}
